+++ hypervisor/include/arch/x86/asm/e820.h
@@ -37,7 +37,7 @@ struct mem_range {
 /* HV read multiboot header to get e820 entries info and calc total RAM info */
 void init_e820(void);
 
-uint64_t e820_alloc_memory(uint32_t size_arg, uint64_t max_addr);
+uint64_t e820_alloc_memory(uint64_t size_arg, uint64_t max_addr);
 /* get total number of the e820 entries */
 uint32_t get_e820_entries_count(void);
